Finish the deploy script for database
Now, the script will use `database/database_deploy.sh` to initialize the
database. There are some configurations for the database in
`config_default.json`. In the `README-zh.md` file, I have added some
instructions for the configurations of the database.

This should solve #8. But there are still some problems:
* When there are tables before, the script will issue an error, but will
not stop.
* There is no clean up for the database.