Fix a bug: use the wrong password in helper
In the helper script, we use the wrong password to login the server.
We used the password for the user in database rather in the OS, which
is wrong. We did not find this bug before because we use the same
password for both the user in database and the user in OS.