Finish configuring mybatis-plus
In this commit, we add the `mybatis-plus` dependency to the `pom.xml`
file, and configure the `mybatis-plus` in the `application-dev.yml`
file. We also add the `UserMapper` interface and `UserPO` class to the
project for testing. The `UserMapper` interface is used to define the
database operations (the real operations will generated by
`mybatis-plus`), and the `UserPO` class is used to define the user
entity. We use `spring-boot-test` to test this function (I've add one
item in the `t_user` table).

This solved #28.