Add filter for authentication and authorization
We add a filter to check the token in the header of the request. If the
token is valid, the request will be passed to the next filter.
Otherwise, the request will be rejected. Besides, we add an
`AuthenticationController` to handle the sign up, sign in, and sign out
requests.

Those features which are not finished yet are marked with `TODO`. We
will finish them ASAP.

This is related with #36. We will use `JJWT` only in this version, maybe
we will use `Spring Security` in the future version.