Add an api for update user information
During development, we found that the `Long` can not be expressed
correctly in the `Swagger` document, so we use `String` to represent the
`id` and convert it to `Long` when we use it as the field of `UserPO`.

We also found that the `HttpServletRequest`'s `getReader()` and
`getInputStream()` can only be called once, so we need to cache the
request body in the `JwtFilter` to make it can be read multiple times.

During this commit, we update the header parameter `Token` with the
`Access-Token` and `Refresh-Token` to make it more clear.

See #32.