Update deploy_helper to append config directly
Before, we'll read the config file, append the new config, and write it
back to the file. This is not ideal, and this caused the file to be too
long sometimes, and it's hard to read. So we update the script to append
the new config directly to the file. In the `application.properties`,
even with the same key, the last one will be used. Therefore, we can
just append the new config to the file.