Modify the tables and add fields
In this submission, the original redundant table `t_user_repository`
was deleted, and the current table `t_user_star_repository` was added
to reflect the collection relationship.

And now two fields are added to the repository table, `https_url` for the
link under the https protocol and `ssh_url` for the link under the ssh
protocol.

See #46 and #48.