Fix a bug related with Sys-Init-V
Now we use a option `serviceType` to control the type of service, the
value can be `systemd` or `sys-init-v`, which means we now can use
`systemd` to deploy in the docker if `systemd` is supported in the
docker. Besides, we remove two unused options. What's more, we update
the script for `sys-init-v`, the old one has some problems: it can not
redirect the output to the log file, all the output was redirected to
the `PIDFILE`, so we just fix this. And, for creating `git` user, there
may be no `/home/git/.ssh`, so we add a `mkdir` command to create it.