Finish the conversion to gitolite
We used to add `ssh-key` to `authorized_keys` simply, which can cause
the problem that all users can access all repositories. Now we use
`gitolite` to manage the access control. This solved #66. After this,
we can support private repository creation.

Besides, we add a new script for developers to prepare the development.