Fix a bug: can not create repo in `gitolite`
This is mainly caused by the `xxx =`. In `gitolite`, it is impossible
to do this. So, now we add `@admin` to collaborators and add `testing`
to every user's repositories. All users can `RW+` the `testing`
repository, but this may not cause the problem, the repository can not
be deleted, because there is no data for this repository in database.

This fixed #99 .