fix(activity): fixed the permission control logic
In this commit,we fixed the permission related to activities and
comments and add some tests. You can consult the topmost comment of
the permissionServiceImpl for details.

Besides,I find a bug that if anyone updates the activity or comment,
it will change userId. It's wrong and we should add an external filed
called modifier_id and change this. I will fix it in this commit.